Reverting types for some functions that were decreasing range of parameters.

+++ SmartDeviceLink/SDLPutFile.h
@@ -33,7 +33,8 @@
 
 - (instancetype)initWithFileName:(NSString *)fileName fileType:(SDLFileType *)fileType persistentFile:(BOOL)persistentFile;
 
-- (instancetype)initWithFileName:(NSString *)fileName fileType:(SDLFileType *)fileType persistentFile:(BOOL)persistentFile systemFile:(BOOL)systemFile offset:(UInt32)offset length:(UInt32)length;
+// TODO: (Alex M.)[2016-12-1] Change from UInt64 ot UInt32
+- (instancetype)initWithFileName:(NSString *)fileName fileType:(SDLFileType *)fileType persistentFile:(BOOL)persistentFile systemFile:(BOOL)systemFile offset:(UInt64)offset length:(UInt64)length;
 
 /**
  * A file reference name
